refactor: Split into lib for use in other components
